voucher/requirements.txt
Flask
supabase
python-dotenv
flask_cors